+++ src/app/feature/admin/components/admin-dashboard/admin-dashboard.component.ts
@@ -1,4 +1,9 @@
 import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
+
+// rxjs
+import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 
 @Component({
   selector: 'app-admin-dashboard',
@@ -6,10 +11,25 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./admin-dashboard.component.css']
 })
 export class AdminDashboardComponent implements OnInit {
-
-  constructor() { }
+  sessionId: Observable<string>;
+  token: Observable<string>;
+  
+  constructor(    private route: ActivatedRoute) { }
 
   ngOnInit() {
+    // Capture the session ID if available
+    this.sessionId = this.route
+      .queryParamMap
+      .pipe(
+        map(params => params.get('sessionId') || 'None')
+      );
+
+    // Capture the fragment if available
+    this.token = this.route
+      .fragment
+      .pipe(
+        map(fragment => fragment || 'None')
+      );
   }
 
 }