+++ src/app/feature/core/guard/auth.guard.ts
@@ -1,0 +1,54 @@
+import { Injectable } from '@angular/core'; import {
+  CanActivate, CanActivateChild, CanLoad, Router, Route,
+  ActivatedRouteSnapshot, RouterStateSnapshot, NavigationExtras, UrlTree, UrlSegment
+} from '@angular/router';
+import { Observable } from 'rxjs';
+import { AuthService } from 'src/app/core/services/auth.service';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class AuthGuard implements CanActivate, CanActivateChild, CanLoad  {
+
+  constructor(
+    private authService: AuthService,
+    private router: Router
+  ) { }
+  canLoad(route: Route, segments: UrlSegment[]): Observable<boolean> | Promise<boolean> | boolean {
+    console.log('CanLoad Guard is called');
+    const url = `/${route.path}`;
+    return <boolean>this.checkLogin(url);
+}
+
+  canActivate(next: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+    console.log('CanActivate Guard is called');
+    const { url } = state;
+    return this.checkLogin(url);
+  }
+  canActivateChild(next: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean | UrlTree> | Promise<boolean | UrlTree> | boolean | UrlTree {
+    console.log('CanActivateChild Guard is called');
+    const { url } = state;
+    return this.checkLogin(url);
+  }
+
+  private checkLogin(url: string): boolean | UrlTree {
+    if (this.authService.isLoggedIn) { return true; }
+
+    // Store the attempted URL for redirecting
+    this.authService.redirectUrl = url;
+    // Create a dummy session id
+    const sessionId = 123456789;
+
+    const navigationExtras: NavigationExtras = {
+      queryParams: { sessionId },
+      fragment: 'anchor'
+    };
+
+    // Navigate to the login page with extras
+    this.router.navigate(['/login'], navigationExtras);
+    return false;
+
+  }
+
+
+}